Move client, redux, server, counters out of modules folder
