+++ packages/client/src/modules/counter/Counter.native.jsx
@@ -2,9 +2,9 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { StyleSheet, View } from 'react-native';
 
-import ClientCounter from './modules/clientCounter/containers/ClientCounter';
-import ServerCounter from './modules/serverCounter/containers/ServerCounter';
-import ReduxCounter from './modules/reduxCounter/containers/ReduxCounter';
+import ClientCounter from './clientCounter/containers/ClientCounter';
+import ServerCounter from './serverCounter/containers/ServerCounter';
+import ReduxCounter from './reduxCounter/containers/ReduxCounter';
 import translate from '../../i18n';
 
 const Counter = ({ t }) => (
