+++ packages/client/src/modules/counter/clientCounter/components/ClientCounterView.native.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { StyleSheet, Text, View } from 'react-native';
 
-import { Button, primary } from '../../../../common/components/native';
+import { Button, primary } from '../../../common/components/native/index';
 
 export const ClientCounterView = ({ text, children }) => (
   <View>
