+++ packages/client/src/modules/counter/clientCounter/containers/ClientCounter.jsx
@@ -5,7 +5,7 @@ import { Mutation, Query } from 'react-apollo';
 import { ClientCounterButton, ClientCounterView } from '../components/ClientCounterView';
 import COUNTER_QUERY_CLIENT from '../graphql/CounterQuery.client.graphql';
 import ADD_COUNTER_CLIENT from '../graphql/AddCounter.client.graphql';
-import translate from '../../../../../i18n';
+import translate from '../../../../i18n/index';
 
 const IncreaseButton = ({ counterAmount, t }) => (
   <Mutation mutation={ADD_COUNTER_CLIENT}>
