+++ packages/client/src/modules/counter/clientCounter/graphql/CounterQuery.client.graphql
