+++ packages/client/src/modules/counter/clientCounter/index.web.jsx
@@ -1,6 +1,6 @@
 import resolvers from './resolvers/index';
 import resources from './locales/index';
-import Feature from '../../../connector';
+import Feature from '../../connector';
 
 export default new Feature({
   resolver: resolvers,
