+++ packages/client/src/modules/counter/clientCounter/locales/en/translations.json
