+++ packages/client/src/modules/counter/clientCounter/locales/index.js
