+++ packages/client/src/modules/counter/reduxCounter/components/ReduxCounterView.native.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { StyleSheet, Text, View } from 'react-native';
 
-import { Button, primary } from '../../../../common/components/native/index';
+import { Button, primary } from '../../../common/components/native/index';
 
 export const ReduxCounterView = ({ text, children }) => (
   <View>
