+++ packages/client/src/modules/counter/reduxCounter/containers/ReduxCounter.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 
 import { ReduxCounterButton, ReduxCounterView } from '../components/ReduxCounterView';
-import translate from '../../../../../i18n';
+import translate from '../../../../i18n/index';
 
 const ReduxCounter = ({ t, onReduxIncrement, reduxCount }) => (
   <ReduxCounterView text={t('text', { reduxCount })}>
