+++ packages/client/src/modules/counter/reduxCounter/index.native.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import { createStackNavigator } from 'react-navigation';
 
-import translate from '../../../../i18n/index';
-import { HeaderTitle } from '../../../common/components/native/index';
-import ClientCounter from './containers/ClientCounter';
-import resolvers from './resolvers/index';
+import translate from '../../../i18n/index';
+import { HeaderTitle } from '../../common/components/native/index';
+import ReduxCounter from './containers/ReduxCounter';
+import reducers from './reducers/index';
 import resources from './locales/index';
-import Feature from '../../../connector';
+import Feature from '../../connector';
 
 const HeaderTitleWithI18n = translate('counter')(HeaderTitle);
 
@@ -15,7 +15,7 @@ export default new Feature({
     Counter: {
       screen: createStackNavigator({
         Counter: {
-          screen: ClientCounter
+          screen: ReduxCounter
         }
       }),
       navigationOptions: {
@@ -23,6 +23,6 @@ export default new Feature({
       }
     }
   },
-  resolver: resolvers,
-  localization: { ns: 'clientCounter', resources }
+  reducer: { counter: reducers },
+  localization: { ns: 'reduxCounter', resources }
 });
