+++ packages/client/src/modules/counter/reduxCounter/index.web.jsx
@@ -1,5 +1,5 @@
 import resources from './locales/index';
-import Feature from '../../../connector';
+import Feature from '../../connector';
 import reducers from './reducers/index';
 
 export default new Feature({
