+++ packages/client/src/modules/counter/reduxCounter/locales/en/translations.json
