+++ packages/client/src/modules/counter/reduxCounter/locales/index.js
