+++ packages/client/src/modules/counter/serverCounter/index.web.jsx
@@ -1,5 +1,5 @@
 import resources from './locales/index';
-import Feature from '../../../connector';
+import Feature from '../../connector';
 
 export default new Feature({
   localization: { ns: 'serverCounter', resources }
