+++ packages/client/src/modules/counter/serverCounter/locales/en/translations.json
