+++ packages/client/src/modules/counter/serverCounter/locales/index.js
